:no_bicycles: Handle no nearest point
The point of handling no nearest point is basically just that I'm
causing a bunch of NaNs messing around with physics and so it doesn't
find any points and I don't want to see irrelevant errors.
But if I want to have a maximum distance for dragging I'll definitely
need this.